+++ aov.mjs
@@ -5,6 +5,7 @@ import renderSceneControls from "./system/hooks/render-scene-controls.mjs";
 import RenderJournalEntrySheet from './system/hooks/render-journal-entry-sheet.mjs'
 import RenderRollTableSheet from './system/hooks/render-roll-table-sheet.mjs'
 import CreateToken from './system/hooks/create-token.mjs'
+import RenderChatMessageHTML from './system/hooks/render-chat-message.mjs'
 
 import Init from './system/hooks/init.mjs';
 import Ready from './system/hooks/ready.mjs';
@@ -17,5 +18,6 @@ Hooks.on('getSceneControlButtons', AOVMenu.getButtons)
 Hooks.on('renderSceneControls', renderSceneControls);
 Hooks.on('renderJournalEntrySheet', RenderJournalEntrySheet);
 Hooks.on('renderRollTableSheet', RenderRollTableSheet);
+Hooks.on('renderChatMessageHTML', RenderChatMessageHTML);
 Hooks.on('createToken', CreateToken);