+++ system/hooks/render-chat-message.mjs
@@ -1,0 +1,6 @@
+/* global $ */
+import { AOVChat } from '../apps/chat.mjs'
+
+export default function (app, html, data) {
+  AOVChat.renderMessageHook(app, html, data)
+}