+++ system/setup/aov-dialog.mjs
@@ -1,0 +1,9 @@
+export default class AOVDialog extends foundry.applications.api.DialogV2 {
+  static DEFAULT_OPTIONS = {
+    classes: ["aov","item"],
+    position: {
+      width: 400,
+      height: "auto",
+    },
+  }
+}