+++ system/setup/update.mjs
@@ -17,7 +17,12 @@ export async function updateWorld({ bypassVersionCheck=false }={}) {
     //Message if current system is less that Version 13.4
     if (foundry.utils.isNewerVersion('13.4', currentVersion ?? '0')) {
       await updateDialog ('systems/aov/templates/updates/update13.4.hbs')  
-    }    
+    } 
+    
+    //Message if current system is less that Version 13.5
+    if (foundry.utils.isNewerVersion('13.5', currentVersion ?? '0')) {
+      await updateDialog ('systems/aov/templates/updates/update13.5.hbs')  
+    } 
 
    await game.settings.set("aov", "systemVersion", targetVersion);
 }